@@ -7,32 +7,34 @@ environment:
   sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
-  cloud_firestore: ^0.16.0
-  cupertino_icons: ^1.0.0
-  enum_to_string: ^1.0.9
-  firebase_core: ^0.7.0
-  firebase_auth: ^0.20.0+1
+  cloud_firestore: ^1.0.0
+  firebase_crashlytics: ^1.0.0
+  firebase_core: ^1.0.0
+  firebase_auth: ^1.0.0
+  firebase_auth_web: ^1.0.0
+  cupertino_icons: ^1.0.2
+  enum_to_string: ^1.0.14
+  flutter_markdown: ^0.6.0
+  flutter_spinkit: ^5.0.0
+  flutter_svg: ^0.21.0-nullsafety.0
+  font_awesome_flutter: ^9.0.0-nullsafety
+  http: ^0.13.0
+  intl: ^0.17.0
+  package_info: ^2.0.0
+  pedantic: ^1.11.0
+  provider: ^5.0.0
+  url_launcher: ^6.0.2
+  syncfusion_flutter_charts: ^18.4.44+1
+  syncfusion_flutter_gauges: ^18.4.44
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_markdown: ^0.5.0
-  flutter_spinkit: ^4.1.2+1
-  flutter_svg: ^0.19.0
-  font_awesome_flutter: ^8.8.1
-  http: ^0.12.0+4
-  intl: ^0.16.1
-  package_info: ^0.4.1
-  pedantic: ^1.8.0
-  provider: ^4.1.2
-  url_launcher: ^5.5.3
-  syncfusion_flutter_charts: ^18.4.44+1
-  syncfusion_flutter_gauges: ^18.4.44
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.2
+  mockito: ^5.0.0
 
 
 flutter:
