@@ -5,24 +5,25 @@ class Dialogs {
       BuildContext context, GlobalKey key, String message) async {
     return showDialog(
         context: context,
-        child: SimpleDialog(
-            key: key,
-            contentPadding: const EdgeInsets.all(24),
-            shape:
-                RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
-            children: <Widget>[
-              Center(
-                child: Column(children: [
-                  CircularProgressIndicator(),
-                  SizedBox(
-                    height: 10,
-                  ),
-                  Text(
-                    message,
-                    style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+        builder: (BuildContext context) => SimpleDialog(
+                key: key,
+                contentPadding: const EdgeInsets.all(24),
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(15)),
+                children: <Widget>[
+                  Center(
+                    child: Column(children: [
+                      CircularProgressIndicator(),
+                      SizedBox(
+                        height: 10,
+                      ),
+                      Text(
+                        message,
+                        style: TextStyle(
+                            fontSize: 20, fontWeight: FontWeight.bold),
+                      )
+                    ]),
                   )
-                ]),
-              )
-            ]));
+                ]));
   }
 }
