@@ -13,7 +13,8 @@ class PlayerWidget extends StatelessWidget {
   Future _showEditPlayerDialog(BuildContext context) async {
     await showDialog(
         context: context,
-        child: PlayerInfoDialog(player: player, isEditing: false));
+        builder: (BuildContext context) =>
+            PlayerInfoDialog(player: player, isEditing: false));
   }
 
   @override
